requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
cssutils==1.0.2
currency-symbols==1.0.0
Django==3.1.3
et-xmlfile==1.0.1
fonttools==4.18.2
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
lxml==4.6.2
openpyxl==3.0.5
Pillow==8.1.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
webcolors==1.11.1
xlrd==2.0.1
xls2xlsx==0.1.5
